Adding spanish_trump.character.json
Many people are not native English speakers & a template was needed to make it easier for anyone to create an agent that speaks other languages by default. 
Modified Trump character file to have the agent speak in Latin Spanish by default. 
This file can serve as a template for creating agents who speak a specific language by default for a given language audience.
The agent will respond in any language if specifically asked to respond in that language in the user input.